internal/glfw: merge Windows/Darwin file pairs into shared files
The files should not contain platform-specific code and were already
nearly identical. Merge them into shared files with a
`//go:build darwin || windows` tag and reconcile the small differences
to pick the correct behavior on both platforms. Add a Windows
`platformCreateCursor` based on the existing `createIcon` helper
(matching GLFW upstream) so `CreateCursor` can live in the shared file.

Closes #3431
